requirements.txt
# 如意 Agent 依赖

# 核心依赖
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0

# 日志
loguru>=0.7.2

# YAML 配置
pyyaml>=6.0.1

# 可选依赖
# psutil>=5.9.0  # 系统信息功能
